# beacon-server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "MeshCore network observation backend. Ingests LoRa packets from MQTT brokers, stores in PostgreSQL, and streams live events via WebSocket.",
        "title": "MeshCore Beacon API",
        "termsOfService": "https://github.com/MeshCore-Beacon/beacon-server",
        "contact": {
            "name": "MeshCore Beacon",
            "url": "https://github.com/MeshCore-Beacon/beacon-server"
        },
        "license": {
            "name": "AGPL-3-or-later"
        },
        "version": "1.6.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/nodes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nodes"
                ],
                "summary": "List nodes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Node type integer (1=companion, 2=repeater, 3=room_server, 4=sensor)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Node type name (companion, repeater, room_server, sensor)",
                        "name": "typeName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by single IATA code (case-insensitive)",
                        "name": "iata",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by multiple IATA codes, comma-separated e.g. YVR,YYJ",
                        "name": "iatas",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by region ID, expands to member IATAs",
                        "name": "regionId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by region slug, expands to member IATAs",
                        "name": "region",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Partial case-insensitive name match",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by transport scope name e.g. %23bc (URL-encoded #bc)",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Exact public key match (hex)",
                        "name": "pubkey",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Partial public key match: hex prefix, case-insensitive",
                        "name": "pubkeyPrefix",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by multibyte path support (true/false); omit for no filter",
                        "name": "supportsMultibytePaths",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by multibyte trace support (true/false); omit for no filter",
                        "name": "supportsMultibyteTraces",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include each node's known neighbor IDs (neighborIds field). Bare ?neighbors or ?neighbors=true enables it; omit/false for none",
                        "name": "neighbors",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field: name, type, radio, neighbors, last_seen (default last_seen)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort direction: asc or desc (default desc)",
                        "name": "direction",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Opaque keyset cursor returned as nextPageToken",
                        "name": "pageToken",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Legacy last_seen epoch ms cursor (only for last_seen desc)",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 50)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_NodeSummary"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
        "/nodes/{nodeId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nodes"
                ],
                "summary": "Get node detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node UUID",
                        "name": "nodeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Node"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
        "/nodes/{nodeId}/neighbors": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nodes"
                ],
                "summary": "List neighbors for a node",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node UUID",
                        "name": "nodeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.NodeNeighbor"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
        "/nodes/{nodeId}/observations": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nodes"
                ],
                "summary": "List packet observations originating from a node",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node UUID",
                        "name": "nodeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Observation ID of last item for pagination",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 50)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_PacketObservationSummary"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.Node": {
            "type": "object",
            "properties": {
                "clockCheckedAt": {
                    "type": "integer"
                },
                "clockDriftSeconds": {
                    "description": "Clock drift, repeaters/room servers only (nodeType 2/3); omitted entirely for other\nnode types or when no qualifying advert has been measured yet. Device minus server\ntime, in seconds, from the advert's self-reported timestamp: +ve = device ahead.\nclockCheckedAt is the server receive time of the advert this was measured from (same\nmoment as lastAdvertAt). clockOutOfSync is |clockDriftSeconds| exceeding a configured\nthreshold (default 5m) -- see internal/config.ResolvedConfig.ClockDriftThreshold.",
                    "type": "integer"
                },
                "clockOutOfSync": {
                    "type": "boolean"
                },
                "defaultScope": {
                    "description": "most recently matched transport scope name e.g. \"#bc\"",
                    "type": "string"
                },
                "firstSeen": {
                    "description": "epoch ms",
                    "type": "integer"
                },
                "iatas": {
                    "description": "IATAs where this node has been heard, with last heard timestamps",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.NodeIATA"
                    }
                },
                "id": {
                    "type": "string"
                },
                "isObserver": {
                    "description": "true if this node is also a known observer",
                    "type": "boolean"
                },
                "knownNeighborCount": {
                    "type": "integer"
                },
                "lastAdvertAt": {
                    "description": "epoch ms, nil if no advert received",
                    "type": "integer"
                },
                "lastSeen": {
                    "description": "epoch ms",
                    "type": "integer"
                },
                "lat": {
                    "description": "decimal degrees, from advert AppData",
                    "type": "number"
                },
                "lng": {
                    "description": "decimal degrees, from advert AppData",
                    "type": "number"
                },
                "locationSource": {
                    "description": "\"advert\" or \"manual\"",
                    "type": "string"
                },
                "metadata": {
                    "description": "raw JSONB metadata"
                },
                "minFirmwareVersion": {
                    "description": "derived from capability flags",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "neighborIds": {
                    "description": "only populated when the list request opts in; see ?neighbors=true",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "neighbors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.NodeNeighbor"
                    }
                },
                "nodeType": {
                    "description": "1=companion, 2=repeater, 3=room_server, 4=sensor",
                    "type": "integer"
                },
                "nodeTypeName": {
                    "type": "string"
                },
                "observerId": {
                    "description": "UUID of the associated observer row, if any",
                    "type": "string"
                },
                "publicKey": {
                    "description": "hex-encoded Ed25519 public key",
                    "type": "string"
                },
                "radio": {
                    "description": "shorthand: \"freqMhz,bwKhz,sf\" e.g. \"910.525,62.5,7\"",
                    "type": "string"
                },
                "stale": {
                    "description": "Stale is true when the node hasn't been seen (last_seen) within the configured\nstaleness window (default 24h; internal/config.ResolvedConfig.NodeStaleThreshold).\nApplies to every node type, unlike ClockDriftSeconds/ClockOutOfSync on Node, which\nare repeater/room-server only.",
                    "type": "boolean"
                },
                "supportsMultibytePaths": {
                    "description": "firmware \u003e= 1.14.0; detected via path hash size",
                    "type": "boolean"
                },
                "supportsMultibyteTraces": {
                    "description": "firmware \u003e= 1.11.0; detected via trace hash size",
                    "type": "boolean"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.NodeIATA": {
            "type": "object",
            "properties": {
                "iata": {
                    "type": "string"
                },
                "lastHeard": {
                    "description": "epoch ms",
                    "type": "integer"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.NodeNeighbor": {
            "type": "object",
            "properties": {
                "firstSeen": {
                    "description": "epoch ms",
                    "type": "integer"
                },
                "iata": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastSeen": {
                    "description": "epoch ms",
                    "type": "integer"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "nodeType": {
                    "type": "integer"
                },
                "nodeTypeName": {
                    "type": "string"
                },
                "observationCount": {
                    "type": "integer"
                },
                "publicKey": {
                    "type": "string"
                },
                "snr": {
                    "type": "number"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.NodeSummary": {
            "type": "object",
            "properties": {
                "defaultScope": {
                    "description": "most recently matched transport scope name e.g. \"#bc\"",
                    "type": "string"
                },
                "iatas": {
                    "description": "IATAs where this node has been heard, with last heard timestamps",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.NodeIATA"
                    }
                },
                "id": {
                    "type": "string"
                },
                "isObserver": {
                    "description": "true if this node is also a known observer",
                    "type": "boolean"
                },
                "knownNeighborCount": {
                    "type": "integer"
                },
                "lat": {
                    "description": "decimal degrees, from advert AppData",
                    "type": "number"
                },
                "lng": {
                    "description": "decimal degrees, from advert AppData",
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "neighborIds": {
                    "description": "only populated when the list request opts in; see ?neighbors=true",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "nodeType": {
                    "description": "1=companion, 2=repeater, 3=room_server, 4=sensor",
                    "type": "integer"
                },
                "nodeTypeName": {
                    "type": "string"
                },
                "observerId": {
                    "description": "UUID of the associated observer row, if any",
                    "type": "string"
                },
                "publicKey": {
                    "description": "hex-encoded Ed25519 public key",
                    "type": "string"
                },
                "radio": {
                    "description": "shorthand: \"freqMhz,bwKhz,sf\" e.g. \"910.525,62.5,7\"",
                    "type": "string"
                },
                "stale": {
                    "description": "Stale is true when the node hasn't been seen (last_seen) within the configured\nstaleness window (default 24h; internal/config.ResolvedConfig.NodeStaleThreshold).\nApplies to every node type, unlike ClockDriftSeconds/ClockOutOfSync on Node, which\nare repeater/room-server only.",
                    "type": "boolean"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.PacketObservationSummary": {
            "type": "object",
            "properties": {
                "heardAt": {
                    "description": "epoch ms",
                    "type": "integer"
                },
                "hopCount": {
                    "type": "integer"
                },
                "iata": {
                    "type": "string"
                },
                "id": {
                    "description": "observation ID, use as cursor for pagination",
                    "type": "integer"
                },
                "packetHash": {
                    "description": "hex-encoded",
                    "type": "string"
                },
                "payloadType": {
                    "type": "integer"
                },
                "payloadTypeName": {
                    "type": "string"
                },
                "rssi": {
                    "type": "integer"
                },
                "snr": {
                    "type": "number"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_NodeSummary": {
            "type": "object",
            "properties": {
                "hasMore": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.NodeSummary"
                    }
                },
                "nextCursor": {
                    "type": "integer"
                },
                "nextPageToken": {
                    "type": "string"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_PacketObservationSummary": {
            "type": "object",
            "properties": {
                "hasMore": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PacketObservationSummary"
                    }
                },
                "nextCursor": {
                    "type": "integer"
                },
                "nextPageToken": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_handlers.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "e.g. \"not_found\", \"bad_request\"",
                    "type": "string"
                },
                "message": {
                    "description": "e.g. \"channel not found\"",
                    "type": "string"
                }
            }
        },
...
    },
...
}